requirements.txt
alembic==1.14.1
amqp==5.3.1
aniso8601==10.0.0
attrs==25.1.0
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.1
Deprecated==1.2.18
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.3.4
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-mock==1.12.1
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
